fix focus after submitting
